fix(profile): include current end user in parties overview search and filter results
